+++ next-docs/public/examples/checkboxTW/Customize.tsx
@@ -1,0 +1,12 @@
+import { Checkbox } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <>
+    <Checkbox bgColor="bg-krillin" />
+    <Checkbox className="rounded-none" />
+    <Checkbox className="shadow-krillin" />
+    <Checkbox className="text-krillin-60" />
+  </>
+);
+
+export default Example;