+++ workspaces/core/src/checkbox/private/types/CheckboxProps.ts
@@ -5,6 +5,8 @@ type CheckboxProps = {
   disabled?: boolean;
   readOnly?: boolean;
   checked?: boolean;
+  bgColor?: string;
+  className?: string;
   onClick?: (event?: React.MouseEvent<HTMLInputElement>) => void;
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
 };