+++ packages/draggabletable/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 9.14.1
+
+### Patch Changes
+
+- Bump patch version
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.14.1
+  - @heathmont/moon-components@9.14.1
+  - @heathmont/moon-table@9.14.1
+  - @heathmont/moon-utils@9.14.1
+
 ## 9.14.0
 
 ### Minor Changes