+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 9.14.1
+
+### Patch Changes
+
+- Bump patch version
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.14.1
+
 ## 9.14.0
 
 ### Minor Changes